Push so @jacksos101 can look at it
cc @jacksos101 I need these relationships fixed.
From a page I should be able to look up all revisions. A revision should be able to look at it's page.

A Revision should be able to look at it's text fields.
A text field it's revisions and links
and a link should be able to look at it's text field/s