File downloads
This PR makes it possible to download media by passing the url query `download=true` to the query string. For example: `/media?filename=somefilename&download=true`.

The downloaded file will have a short name.

I've also added a title proerty for the select media modal. If no title is provided then the title of the modal will be `select an {this.props.type}`